requirements.txt
pandas>=1.5.0
pyperclip>=1.8.0
pyyaml>=6.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0